Implemented the Initial Blog Functionality + Added GH Language Breakdown to About Page! (#2)
* initializing create_blog branch

* reorganized component organization and centered blog posts

* changed location of gitignore

* flask dynamically updates to changes

* Blog entry data recieved by api and sqlite database started

* Blog table not found :(

* database updated within docker container and file copied to personal machine

* database file does not dynamically update

* non-working changes

* posts populated by database

* database insertions infinite looping

* database still not working, made sure blog posts only show up once after mounting

* unworking changes. Post param changes

* stored to database correctly. not displaying correctly

* title/content not displayed in post

* posts are correctly populated, put page refresh needed to load new posts

* post list updated and page reacts when new post submitted!!!

* add likes and time formatting today

* messing around with styling

* adding language stats to about me page

* correcly calculating github lines, but backend breaking in docker compose

* fixed compose issue

* added shadcn.ui component library!!!

* bar changes colors according to percentage

* mimic gh language breakdown

* removed percentages and total from displaying and corrected rounding

* adding likes and update statement

* update saved to db

* reformatting db calls

* reformatting almost done, insert working, UPDATE IS NOT

* likes successfully updated

* trying to implement cache for getting languages

* almost done implementing cache for getting use languages

* almost working :)

* error when trying to insert strings w/ special characters into db

* unworking changes, able to read from db. make sure to make the languages table store json strings

* json dumps/loads implemented, special characters still dont work

* DATABASE CAN HANDLE SPECIAL CHARACTERS!!!

* db WORKING and posts update to correct size!

* update blog to be my LOCAL TIME DATE

* updated timezone

+++ Makefile
@@ -17,10 +17,18 @@ watch:
 stack: 
 	@echo "Building stack"
 	(make watch &)
-	docker-compose up
+	./compose.sh
+
+get-db:
+	@echo "Copying database file from container to host..."
+	docker cp personal-portfolio-backend-1:/app/backend/database/portfolio.db $$(bash -c "pwd")/backend/backend/database/portfolio.db
 
 clean:
 	@echo "Cleaning up"
-	docker stop $(shell docker ps -aq)
-	docker rm $(shell docker ps -aq)
-	docker rmi $(shell docker images -aq)
+	if [ "`docker ps -aq`" ]; then docker stop $(shell docker ps -aq); fi
+	if [ "`docker ps -aq`" ]; then docker rm $(shell docker ps -aq); fi
+	if [ "`docker images -aq`" ]; then docker rmi $(shell docker images -aq); fi
+
+format:
+	(cd frontend && yarn format)
+	black .
